pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "logsift"
version = "0.1.0"
description = "Format-aware CLI for searching and filtering structured logs"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "farkhad" }]
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "python-dateutil>=2.8",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
logsift = "logsift.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/logsift"]
